backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
python-dotenv
alembic
flask-jwt-extended
flask-mail
requests
gunicorn
Flask-Limiter
uuid6
